MAINT: Rename xoshiro256/512starstar
Drop the starstar since these are the preferred generators
Fix table in docs with jumped info
