+++ doc/source/reference/random/bit_generators/index.rst
@@ -24,8 +24,8 @@ Stable RNGs
    ThreeFry <threefry>
    XoroShiro128+ <xoroshiro128>
    Xorshift1024*φ <xorshift1024>
-   Xoshiro256** <xoshiro256starstar>
-   Xoshiro512** <xoshiro512starstar>
+   Xoshiro256** <xoshiro256>
+   Xoshiro512** <xoshiro512>
 
 
 Experimental RNGs
