+++ doc/source/reference/random/parallel.rst
@@ -75,6 +75,10 @@ and the bits in the default unsigned random are listed below.
 +-----------------+-------------------------+-------------------------+-------------------------+
 | Xorshift1024    | :math:`2^{1024}`        | :math:`2^{512}`         | 64                      |
 +-----------------+-------------------------+-------------------------+-------------------------+
+| Xoshiro256**    | :math:`2^{256}`         | :math:`2^{128}`         | 64                      |
++-----------------+-------------------------+-------------------------+-------------------------+
+| Xoshiro512**    | :math:`2^{512}`         | :math:`2^{256}`         | 64                      |
++-----------------+-------------------------+-------------------------+-------------------------+
 
 ``jumped`` can be used to produce long blocks which should be long enough to not
 overlap.
