+++ numpy/random/src/xoshiro256/xoshiro256.orig.c
