+++ numpy/random/src/xoshiro512/LICENSE.md
