+++ numpy/random/src/xoshiro512/xoshiro512.c
@@ -6,7 +6,7 @@ worldwide. This software is distributed without any warranty.
 
 See <http://creativecommons.org/publicdomain/zero/1.0/>. */
 
-#include "xoshiro512starstar.orig.h"
+#include "xoshiro512.h"
 
 /* This is xoshiro512** 1.0, an all-purpose, rock-solid generator. It has
    excellent (about 1ns) speed, an increased state (512 bits) that is
@@ -19,49 +19,34 @@ See <http://creativecommons.org/publicdomain/zero/1.0/>. */
    a 64-bit seed, we suggest to seed a splitmix64 generator and use its
    output to fill s. */
 
-static inline uint64_t rotl(const uint64_t x, int k) {
-	return (x << k) | (x >> (64 - k));
-}
+extern NPY_INLINE uint64_t xoshiro512_next64(xoshiro512_state *state);
 
-
-uint64_t next(void) {
-	const uint64_t result_starstar = rotl(s[1] * 5, 7) * 9;
-
-	const uint64_t t = s[1] << 11;
-
-	s[2] ^= s[0];
-	s[5] ^= s[1];
-	s[1] ^= s[2];
-	s[7] ^= s[3];
-	s[3] ^= s[4];
-	s[4] ^= s[5];
-	s[0] ^= s[6];
-	s[6] ^= s[7];
-
-	s[6] ^= t;
-
-	s[7] = rotl(s[7], 21);
-
-	return result_starstar;
-}
+extern NPY_INLINE uint32_t xoshiro512_next32(xoshiro512_state *state);
 
 
 /* This is the jump function for the generator. It is equivalent
    to 2^256 calls to next(); it can be used to generate 2^256
    non-overlapping subsequences for parallel computations. */
 
-void jump(void) {
-	static const uint64_t JUMP[] = { 0x33ed89b6e7a353f9, 0x760083d7955323be, 0x2837f2fbb5f22fae, 0x4b8c5674d309511c, 0xb11ac47a7ba28c25, 0xf1be7667092bcc1c, 0x53851efdb6df0aaf, 0x1ebbc8b23eaf25db };
+static uint64_t s_placeholder[8];
 
-	uint64_t t[sizeof s / sizeof *s];
-	memset(t, 0, sizeof t);
-	for(int i = 0; i < sizeof JUMP / sizeof *JUMP; i++)
-		for(int b = 0; b < 64; b++) {
-			if (JUMP[i] & UINT64_C(1) << b)
-				for(int w = 0; w < sizeof s / sizeof *s; w++)
-					t[w] ^= s[w];
-			next();
-		}
+void xoshiro512_jump(xoshiro512_state *state) {
 
-	memcpy(s, t, sizeof s);
+  int i, b, w;
+  static const uint64_t JUMP[] = {0x33ed89b6e7a353f9, 0x760083d7955323be,
+                                  0x2837f2fbb5f22fae, 0x4b8c5674d309511c,
+                                  0xb11ac47a7ba28c25, 0xf1be7667092bcc1c,
+                                  0x53851efdb6df0aaf, 0x1ebbc8b23eaf25db};
+
+  uint64_t t[sizeof s_placeholder / sizeof *s_placeholder];
+  memset(t, 0, sizeof t);
+  for (i = 0; i < sizeof JUMP / sizeof *JUMP; i++)
+    for (b = 0; b < 64; b++) {
+      if (JUMP[i] & UINT64_C(1) << b)
+        for (w = 0; w < sizeof s_placeholder / sizeof *s_placeholder; w++)
+          t[w] ^= state->s[w];
+      xoshiro512_next(&state->s[0]);
+    }
+
+  memcpy(state->s, t, sizeof s_placeholder);
 }
