+++ numpy/random/src/xoshiro512/xoshiro512.h
@@ -5,19 +5,19 @@
 #include <inttypes.h>
 #include "numpy/npy_common.h"
 
-typedef struct s_xoshiro512starstar_state
+typedef struct s_xoshiro512_state
 {
 	uint64_t s[8];
 	int has_uint32;
 	uint32_t uinteger;
-} xoshiro512starstar_state;
+} xoshiro512_state;
 
 static NPY_INLINE uint64_t rotl(const uint64_t x, int k)
 {
 	return (x << k) | (x >> (64 - k));
 }
 
-static NPY_INLINE uint64_t xoshiro512starstar_next(uint64_t *s)
+static NPY_INLINE uint64_t xoshiro512_next(uint64_t *s)
 {
 	const uint64_t result_starstar = rotl(s[1] * 5, 7) * 9;
 
@@ -39,14 +39,13 @@ static NPY_INLINE uint64_t xoshiro512starstar_next(uint64_t *s)
 	return result_starstar;
 }
 
-static NPY_INLINE uint64_t
-xoshiro512starstar_next64(xoshiro512starstar_state *state)
+
+static NPY_INLINE uint64_t xoshiro512_next64(xoshiro512_state *state)
 {
-	return xoshiro512starstar_next(&state->s[0]);
+	return xoshiro512_next(&state->s[0]);
 }
 
-static NPY_INLINE uint32_t
-xoshiro512starstar_next32(xoshiro512starstar_state *state)
+static NPY_INLINE uint32_t xoshiro512_next32(xoshiro512_state *state)
 {
 	uint64_t next;
 	if (state->has_uint32)
@@ -54,12 +53,12 @@ xoshiro512starstar_next32(xoshiro512starstar_state *state)
 		state->has_uint32 = 0;
 		return state->uinteger;
 	}
-	next = xoshiro512starstar_next(&state->s[0]);
+	next = xoshiro512_next(&state->s[0]);
 	state->has_uint32 = 1;
 	state->uinteger = (uint32_t)(next >> 32);
 	return (uint32_t)(next & 0xffffffff);
 }
 
-void xoshiro512starstar_jump(xoshiro512starstar_state *state);
+void xoshiro512_jump(xoshiro512_state *state);
 
 #endif
