+++ numpy/random/tests/test_randomstate.py
@@ -9,7 +9,7 @@ from numpy.testing import (
         suppress_warnings
         )
 
-from numpy.random import MT19937, Xoshiro256StarStar, mtrand as random
+from numpy.random import MT19937, Xoshiro256, mtrand as random
 
 
 def assert_mt19937_state_equal(a, b):
@@ -166,13 +166,13 @@ class TestSetState(object):
         self.random_state.negative_binomial(0.5, 0.5)
 
     def test_get_state_warning(self):
-        rs = random.RandomState(Xoshiro256StarStar())
+        rs = random.RandomState(Xoshiro256())
         with suppress_warnings() as sup:
             w = sup.record(RuntimeWarning)
             state = rs.get_state()
             assert_(len(w) == 1)
             assert isinstance(state, dict)
-            assert state['bit_generator'] == 'Xoshiro256StarStar'
+            assert state['bit_generator'] == 'Xoshiro256'
 
     def test_invalid_legacy_state_setting(self):
         state = self.random_state.get_state()
